chore: restore new openapi-types file structure
